No compilation errors

+++ hw2/src/main.c
@@ -1,6 +1,5 @@
 #include "debug.h"
 #include "utf.h"
-#include "wrappers.h"
 #include <stdlib.h>
 
 int
@@ -17,10 +16,10 @@ main(int argc, char *argv[])
   lseek(SEEK_SET, program_state->bom_length, infile); /* Discard BOM */
   get_encoding_function()(infile, outfile);
   if(program_state != NULL) {
-    close((int)program_state);
+    close((long)program_state);
   }
   //I think this is how this works
-  free((void*)outfile);
-  free((void*)infile);
+  free((void*)((long)outfile));
+  free((void*)((long)infile));
   return EXIT_SUCCESS;
 }
